+++ packages/get-changelog-cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0
+
+### Feature
+
+-   Add `-r` option to choose reporter
+-   Add `console-jira` reporter
+
 ## 1.0.0
 
 -   Initial release